(feat|COS-4254): Add filtering to contract columns: - Renewal Date - ARR Forecast - Health - Owner - Period
