+++ packages/apps/frontera/src/routes/organizations/src/components/Columns/contracts/Cells/period/PeriodCell.tsx
@@ -1,22 +1,8 @@
+import { getCommittedPeriodLabel } from '@shared/util/committedPeriodLabel';
+
 interface PeriodCellProps {
   className?: string;
   committedPeriodInMonths?: string | number;
-}
-
-export function getCommittedPeriodLabel(months: string | number) {
-  if (`${months}` === '1') {
-    return 'Monthly';
-  }
-
-  if (`${months}` === '3') {
-    return 'Quarterly';
-  }
-
-  if (`${months}` === '12') {
-    return 'Annually';
-  }
-
-  return `${months}-monthly`;
 }
 
 export const PeriodCell = ({ committedPeriodInMonths }: PeriodCellProps) => {
