+++ packages/apps/frontera/src/routes/organizations/src/components/Columns/contracts/Filters/index.ts
@@ -3,3 +3,4 @@ export * from './CurrencyFilter';
 export * from './StatusFilter';
 export * from './RenewalFilter';
 export * from './LtvFilter';
+export * from './PeriodFilter';
