--- packages/apps/frontera/src/routes/organizations/src/components/Columns/organizations/Filters/Forecast/index.ts
@@ -1,1 +1,0 @@
-export * from './ForecastFilter';
