--- packages/apps/frontera/src/routes/organizations/src/components/Columns/organizations/Filters/Owner/index.ts
@@ -1,1 +1,0 @@
-export * from './OwnerFilter';
