--- packages/apps/frontera/src/routes/organizations/src/components/Columns/organizations/Filters/RenewalLikelihood/index.ts
@@ -1,1 +1,0 @@
-export * from './RenewalLikelihoodFilter';
