+++ packages/apps/frontera/src/routes/organizations/src/components/Columns/shared/Filters/Owner/index.ts
@@ -1,0 +1,1 @@
+export * from './OwnerFilter.tsx';
