+++ packages/apps/frontera/src/routes/src/util/committedPeriodLabel.ts
@@ -1,0 +1,15 @@
+export function getCommittedPeriodLabel(months: string | number) {
+  if (`${months}` === '1') {
+    return 'Monthly';
+  }
+
+  if (`${months}` === '3') {
+    return 'Quarterly';
+  }
+
+  if (`${months}` === '12') {
+    return 'Annually';
+  }
+
+  return `${months}-monthly`;
+}
